Homework_6_fixed

+++ Homework_6/Task_2/main.py
@@ -1,4 +1,4 @@
 # Task 2: Rectangle area
-length = float(input("Введите длину прямоугольника: "))
-width = float(input("Введите ширину прямоугольника: "))
+length = int(input("Введите длину прямоугольника: "))
+width = int(input("Введите ширину прямоугольника: "))
 print(f"Площадь прямоугольника: {length * width}") # f-string
